backend/requirements.txt
Django==4.2.3  # Adjust the version to match your Django version
djangorestframework==3.14.0  # For 'rest_framework'
djangorestframework-simplejwt==5.2.2  # For 'rest_framework_simplejwt'
djangorestframework-authtoken==3.14.0  # For 'rest_framework.authtoken'
django-filter==23.2  # For 'django_filters'
djoser==2.1.0  # For 'djoser'
django-cors-headers==3.14.0  # For 'corsheaders'
Pillow==10.0.0  # For image processing